fix: second layer of cmake-extended failures
Each fix in the previous commit uncovered the next problem in the same job;
this workflow has evidently not been run to completion in a long time.

Cross-compile for armhf, past the -mtune error, now hit gpu_cfr_adapter.c
casting a void* straight to uint64_t. Pointers are 32-bit there, so GCC
rejects it as a cast between different sizes. Routed through uintptr_t.

macOS GCC, past deterministic_benchmark.c, now hit badugi_eval.c with
"specified bound 128 exceeds destination size 80". That one is a real
overflow risk rather than a pedantic warning: BadugiHandVal_toString() passes
POKER_EVAL_HANDVAL_STRING_MAX (128) as the buffer length, its own header says
callers must supply that much, and four call sites — including
BadugiHandVal_print() inside the library — hand it an 80-byte buffer. They now
use BadugiHandVal_toString_n() with sizeof(), which is correct whatever the
buffer size.

Linux GCC Debug Coverage, past test_range_equity, now failed
test_performance_regression_detection — a regression I introduced. That case
feeds synthetic numbers to performance_within_tolerance() and asserts it flags
a -60% drop, so gating the function itself broke it. The function is pure
again; the four verdicts on genuinely measured throughput are gated instead,
through PERF_ASSERT.

Create packages got past the absolute install path and reached the real
failure: the TGZ is produced, then the DEB generator aborts with "Debian
package requires a maintainer for a package". CPACK_PACKAGE_CONTACT is now
set.

Also included: equity_hi and equity_lo were structurally always zero on the
range path, since weightedAggregationAccumulate never carried the
nsharehi/nsharelo histograms that fill_result_legacy derives them from. They
are now accumulated, merged and finalized like the other counters.
distribute_probabilities() had a fixed double[8] scratch while already taking
an arbitrary len; the share histogram needs num_players+1 buckets, so it is
sized to ENUM_MAXPLAYERS+1 with a guard.

That exposed a wrong assertion in test_range_equity: it asserted
equity == 0.5*hi + 0.5*lo, which presumes a qualifying low always exists. In
this matchup low shares sum to 0.58 — 42% of boards produce no low and high
takes the whole pot — so that model accounts for only 79% of the pot and
cannot hold. Replaced with invariants that do: equity sums to 1, high shares
sum to 1, low shares sum to the frequency of a low qualifying, and QdJdTc9c,
holding no card at or below eight, never makes one.

+++ src/gpu/gpu_cfr_adapter.c
@@ -12,6 +12,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 
 #ifdef ENABLE_CUDA
 extern int gpu_cfr_cuda_download_state(gpu_cfr_context_t * ctx, cfr_matrix_storage_t * storage);
@@ -245,10 +246,10 @@ int gpu_cfr_adapter_compute_deltas(gpu_cfr_adapter_t *adapter)
     for (int sample = 0; sample < adapter->mc_samples; sample++)
     {
         /* Sample for player 0 */
-        cfr_traverse(adapter, (uint64_t)adapter->game->initial_state, 0, 1.0, 1.0);
+        cfr_traverse(adapter, (uint64_t)(uintptr_t)adapter->game->initial_state, 0, 1.0, 1.0);
 
         /* Sample for player 1 */
-        cfr_traverse(adapter, (uint64_t)adapter->game->initial_state, 1, 1.0, 1.0);
+        cfr_traverse(adapter, (uint64_t)(uintptr_t)adapter->game->initial_state, 1, 1.0, 1.0);
     }
 
     /* Average deltas by visit count */
